// frontend/src/routes/(auth)/signup/signup-schema.ts
import { z } from "zod";
 
export const signupSchema = z.object({
  firstName: z.string().min(2, { message: "First name must be at least 2 characters long" }),
  lastName: z.string().min(2, { message: "Last name must be at least 2 characters long" }),
  email: z.string().email({ message: "Invalid email address" }),
  password: z.string().min(6, { message: "Password must be at least 6 characters long" }),
});
 
export type SignupSchema = typeof signupSchema;